디테일 뷰 컨트롤러 레이아웃 구성 검색한 gif이미지 중 선택 시 선택한 gif와 타이틀 정보를 디테일뷰에 뿌려주기
