--- MovingPictureApp/Controller/GifMain+CollectionView.swift
@@ -1,24 +1,0 @@
-//
-//  GifMain+CollectionView.swift
-//  MovingPictureApp
-//
-//  Created by LeeJaeHyeok on 2022/05/01.
-//
-
-import UIKit
-
-extension GifMainController: UICollectionViewDelegateFlowLayout, UICollectionViewDataSource {
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return 25
-    }
-
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: GifMainCell.identifier, for: indexPath) as? GifMainCell ?? GifMainCell()
-        cell.backgroundColor = .systemGray5
-        return cell
-    }
-
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        return CGSize(width: view.frame.width / 3 - 8, height: 100)
-    }
-}
