--- MovingPictureApp/View/BaseCollectionViewCell.swift
@@ -1,33 +1,0 @@
-//
-//  BaseCollectionViewCell.swift
-//  MovingPictureApp
-//
-//  Created by LeeJaeHyeok on 2022/05/01.
-//
-
-import UIKit
-
-class BaseCollectionViewCell: UICollectionViewCell {
-
-    // MARK: - Properties
-
-    // MARK: - Lifecycle
-
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        configureUI()
-        configureConstraints()
-    }
-
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    func configureUI() {
-
-    }
-
-    func configureConstraints() {
-
-    }
-}
