--- MovingPictureApp/View/BaseView.swift
@@ -1,36 +1,0 @@
-//
-//  BaseView.swift
-//  MovingPictureApp
-//
-//  Created by LeeJaeHyeok on 2022/05/01.
-//
-
-import UIKit
-
-class BaseView: UIView {
-
-    // MARK: - Properties
-
-    // MARK: - Lifecycle
-
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        configureUI()
-        configureConstraints()
-    }
-
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    // MARK: - Helpers
-
-    func configureUI() {
-        backgroundColor = .secondarySystemBackground
-        configureConstraints()
-    }
-
-    func configureConstraints() {
-
-    }
-}
